requirements.txt
# Core
massive
fastapi
uvicorn[standard]
pydantic
python-dotenv
httpx
numpy
pandas
 
# AI
anthropic
groq
 
# Utilities
pytz
tqdm
orjson
pytest